Исправлен ParticleCPUTexture шейдер Исправлен ParticleCPU шейдер Изменена концепция текстур Добавлена функция Utils::SetupDebugOpenGL для записи сообщений от OpenGL Исправлен ParticleRender Исправлен метод SetTexture2D из Shader Удалён метод SetTextureHandle из Shader AttachShadowsData возвразает количество использованных текстурных блоков Исправлена работа логики метода Texture::SubmitData

+++ include/Render/Render.h
@@ -160,7 +160,7 @@ struct DrawData {
 using DrawObject = RenderObject<DrawData>;
 
 struct SamplerData {
-    u64 Handle;
+    u32 Handle;
     TexType TextureType;
     TexDataType TextureDataType;
 };
